pyproject.toml
[tool.poetry]
name = "desmo-api"
version = "0.1.0"
description = ""
authors = ["Panu Oksiala <panu@oksiala.fi>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9.18,<3.13"
ansible-runner = "^2.3.4"
tenacity = "^8.2.3"
python-statemachine = "^2.1.2"
fastapi = "^0.104.1"
aiohttp = {extras = ["speedups"], version = "^3.9.0"}
aiodns = "^3.1.1"
argon2-cffi = "^23.1.0"
ansible = "^8.0.0"
asyncpg = "^0.29.0"


[tool.poetry.group.dev.dependencies]
pydot = "^1.4.2"
mypy = "^1.7.0"
uvicorn = {extras = ["standard"], version = "^0.24.0.post1"}
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
